AndroidPopularLibs: add injection of image services with dagger
